Add special planet record and bytes32 builder and fix the bug
---

Original commit (chain-contracts): b901ae04bec75179e80645cb5ef832e6d7117c9b

+++ chain-contracts/contracts/controllers/modules/UserSpecialPlanetControllable.sol
@@ -144,9 +144,9 @@ contract UserSpecialPlanetControllable is TimeGettable {
     ) << USER_SPECIAL_PLANET_PERMANENCE_RANKUPED_AT_START_BIT) | (bytes32(
       r.createdAt
     ) << USER_SPECIAL_PLANET_PERMANENCE_CREATED_AT_START_BIT) | (bytes32(
-      r.axialCoordinateQ
+      uint16(r.axialCoordinateQ)
     ) << USER_SPECIAL_PLANET_PERMANENCE_AXIAL_COORDINATE_Q_START_BIT) | (bytes32(
-      r.axialCoordinateR
+      uint16(r.axialCoordinateR)
     ) << USER_SPECIAL_PLANET_PERMANENCE_AXIAL_COORDINATE_R_START_BIT);
   }
